apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==11.2.1
opencv-python==4.8.1.78
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1 